Moved enums in fcr_enum.p
